+++ src/api/l_graphics_pass.c
@@ -666,8 +666,10 @@ static int l_lovrPassSetScissor(lua_State* L) {
     lovrPassSetScissor(pass, NULL);
   } else {
     uint32_t scissor[4];
-    scissor[0] = luax_checku32(L, 2);
-    scissor[1] = luax_checku32(L, 3);
+    lua_Integer x = luaL_checkinteger(L, 2);
+    lua_Integer y = luaL_checkinteger(L, 3);
+    scissor[0] = (uint32_t) MAX(x, 0);
+    scissor[1] = (uint32_t) MAX(y, 0);
     scissor[2] = luax_checku32(L, 4);
     scissor[3] = luax_checku32(L, 5);
     luax_assert(L, lovrPassSetScissor(pass, scissor));